fraudfinder/datagen/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
decorator==5.1.0
dill==0.3.4
flask-socketio==5.1.1
Flask==2.0.2
fsspec==2021.10.0
gcsfs==2021.10.0
google-api-core==2.1.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.28.1
google-cloud-bigquery-storage==2.9.1
google-cloud-core==2.1.0
google-cloud-pubsub
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
gunicorn==20.1.0
idna==3.2
idna-ssl==1.1.0
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
libcst==0.3.21
MarkupSafe==2.0.1
multidict==5.2.0
mypy-extensions==0.4.3
numpy==1.19.5
oauthlib==3.1.1
packaging==21.0
pandarallel==1.5.3
pandas==1.1.5
proto-plus==1.19.4
protobuf==3.18.1
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
simple-websocket
six==1.16.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.0
zipp==3.6.0
